% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{bars}
\alias{bars}
\title{Create bar charts for MM items}
\usage{
bars(variable, scale_ = "mm_likert_scale")
}
\arguments{
\item{variable}{variable to chart}

\item{scale_}{scale used to chart variable defined in df_definitions}
}
\description{
Create bar charts for MM items
}
\examples{
bars("sp_plan")
}
