% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{dbl_output}
\alias{dbl_output}
\title{Plots bars and tiles side by side with title centered above both plots}
\usage{
dbl_output(variable)
}
\arguments{
\item{variable}{variable to be applied to both bar and tile plots}
}
\description{
Plots bars and tiles side by side with title centered above both plots
}
\examples{
dbl_output("sp_plan")
}
