% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{mm_summary_graphs}
\alias{mm_summary_graphs}
\title{Create Summary Graphs of mean & sd for variables}
\usage{
mm_summary_graphs(
  variables,
  mean_or_sd = "mean",
  scale_ = "mm_likert_scale",
  wrap_ = 30
)
}
\arguments{
\item{variables}{variable names from df_responses dataframe ex: c("sp_plan", "sp_process")}

\item{mean_or_sd}{Is the summary displaying mean or standard deviation of scores? "mean" or "sd"}

\item{scale_}{What scale is to be used as defined in 'df_definitions' dataframe}

\item{wrap_}{length of characters before wrapping the axis labels}
}
\description{
Create Summary Graphs of mean & sd for variables
}
\examples{
mm_summary_graphs(c("sp_plan", "sp_process"), "sd", "mm_likert_scale", wrap_ = 30)
mm_summary_graphs(c("sp_plan", "sp_process"))
}
