% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{mm_textTable}
\alias{mm_textTable}
\title{Write tables of text}
\usage{
mm_textTable(
  variables,
  names_ = variables,
  collapse_columns = 1,
  arrange_columns = 1,
  page_rows = 20
)
}
\arguments{
\item{variables}{variables from 'df_responses' data to include in the table single or list. Order given will be order of table variabales}

\item{names_}{human readable names for the table headers single or list}

\item{collapse_columns}{numeric vector of columns to collapse ex: 1, 1:2, c(1, 2, 4)}

\item{page_rows}{number of rows to print on each page per table.}

\item{arrange_column}{numeric vector of columns by which to arrange ex: 1, 1:2, c(1, 2, 4)}
}
\value{
outputs kable table
}
\description{
Write tables of text
}
\examples{
mm_textTable(c('participant_role', 'participant_name'), c("Role", "Name"), 1)
}
