% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{mm_theme}
\alias{mm_theme}
\title{Theme for ggplot graphical outputs. Styles are overridden by calls from individual functions.}
\usage{
mm_theme()
}
\description{
Theme for ggplot graphical outputs. Styles are overridden by calls from individual functions.
}
