% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{pivot_table_by}
\alias{pivot_table_by}
\title{SWOT Tables}
\usage{
pivot_table_by(pivot_value, value_name, roles_id)
}
\arguments{
\item{pivot_value}{the naming convention that will be searched for in varibale names for conversion from wide to long format}

\item{value_name}{the name of the specfic value for the table to be generated}

\item{roles_id}{a definitions id value that indexes the order of the organization roles i.e. "chamber_roles"}
}
\description{
SWOT Tables
}
