% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{readin}
\alias{readin}
\title{Read in JSON file based on path to file}
\usage{
readin(
 
    path = "/home/kent/Contracts/missionmet/mm-assessor/survey_monkey/results-example/MM Checkup for Chambers.json"
)
}
\arguments{
\item{path}{the filepath to the JSON file}
}
\value{

}
\description{
Read in JSON file based on path to file
}
\examples{
dataframe <- readin("home/Ricky/tempfile.json")
}
