% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{swot_by_role}
\alias{swot_by_role}
\title{SWOT Tables}
\usage{
swot_by_role(swot_cat_)
}
\arguments{
\item{swot_cat_}{category of the swot item. strengths, weaknesses, opportunities, threats. Will search for the name in the response variable names.}
}
\description{
SWOT Tables
}
