% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{tiles}
\alias{tiles}
\title{Create heatmap 'tile' charts for MM items}
\usage{
tiles(variable, scale_ = "mm_likert_scale")
}
\arguments{
\item{variable}{variable to chart}

\item{scale_}{scale used to chart variable defined in df_definitions}
}
\description{
Create heatmap 'tile' charts for MM items
}
\examples{
tiles("sp_plan")
}
