% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmassessment.R
\name{mm_textTable}
\alias{mm_textTable}
\title{Write tables of text}
\usage{
mm_textTable(variables, names_, collapse_columns)
}
\arguments{
\item{variables}{variables from 'df_responses' data to include in the table single or list}

\item{collapse_columns}{numeric vector of columns to collapse ex: 1, 1:2, c(1, 2, 4)}

\item{names}{human readable names for the table headers single or list}
}
\value{
outputs kable table
}
\description{
Write tables of text
}
\examples{
mm_textTable(c('participant_role', 'participant_name'), c("Role", "Name"), 1)
}
